Separated login and registration into individual components
